requirements.txt
azure-functions==1.23.0
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
MarkupSafe==3.0.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
python-multipart==0.0.6
pyzbar==0.1.9
requests==2.32.4
requests-toolbelt==1.0.0
urllib3==2.5.0
Werkzeug==3.1.3
